requirements.txt
django-debug-toolbar
Django
